# docs/swagger.yaml
basePath: /
definitions:
  common.ErrorResponse:
    properties:
      details:
        additionalProperties: true
        type: object
      error_code:
        type: string
      message:
        type: string
    type: object
...
  httphandlers.AnalysisLaunch:
    properties:
      app_description:
        type: string
      app_id:
        type: string
      app_name:
        type: string
      archive_logs:
        type: boolean
      batch_id:
        type: string
      condor_id:
        type: string
      condor_log_path:
        description: comes from config, not upstream service
        type: string
      config_file:
        description: path to the job configuration file (not from upstream)
        type: string
      create_output_subdir:
        type: boolean
      date_completed:
        type: string
      date_started:
        type: string
      date_submitted:
        type: string
      description:
        type: string
      email:
        type: string
      execution_target:
        type: string
      exit_code:
        type: integer
      extra:
        $ref: '#/definitions/model.ExtraInfo'
      failure_count:
        type: integer
      failure_threshold:
        type: integer
      file-metadata:
        items:
          $ref: '#/definitions/model.FileMetadata'
        type: array
      filter_files:
        description: comes from config, not upstream service
        items:
          type: string
        type: array
      group:
        description: untested for now
        type: string
      id:
        type: string
      input_path_list:
        description: path to a list of inputs (not from upstream).
        type: string
      input_ticket_list:
        description: path to a list of inputs with tickets (not from upstream).
        type: string
      irods_base:
        type: string
      mount_data_store:
        type: boolean
      name:
        type: string
      nfs_base:
        type: string
      notify:
        type: boolean
      now_date:
        type: string
      output_dir:
        description: the value parsed out of the JSON. Use OutputDirectory() instead.
        type: string
      output_dir_ticket:
        description: the write ticket for output_dir (assumes output_dir is set correctly).
        type: string
      output_ticket_list:
        description: path to the file of the output dest with ticket (not from upstream).
        type: string
      request_type:
        type: string
      run-on-nfs:
        type: boolean
      skip-parent-meta:
        type: boolean
      steps:
        items:
          $ref: '#/definitions/model.Step'
        type: array
      submission_date:
        type: string
      type:
        type: string
      user_groups:
        items:
          type: string
        type: array
      user_home:
        type: string
      user_id:
        type: string
      username:
        type: string
      uuid:
        type: string
      wiki_url:
        type: string
    type: object
  httphandlers.AsyncData:
    properties:
      analysisID:
        type: string
      ipAddr:
        type: string
      subdomain:
        type: string
    type: object
...
  httphandlers.uuidBody:
    properties:
      uuid:
        type: string
    type: object
  incluster.ConfigMapInfo:
    properties:
      analysisName:
        type: string
      appID:
        type: string
      appName:
        type: string
      creationTimestamp:
        type: string
      data:
        additionalProperties:
          type: string
        type: object
      externalID:
        type: string
      name:
        type: string
      namespace:
        type: string
      userID:
        type: string
      username:
        type: string
    type: object
  incluster.DeploymentInfo:
    properties:
      analysisName:
        type: string
      appID:
        type: string
      appName:
        type: string
      command:
        items:
          type: string
        type: array
      creationTimestamp:
        type: string
      externalID:
        type: string
      group:
        type: integer
      image:
        type: string
      name:
        type: string
      namespace:
        type: string
      port:
        type: integer
      user:
        type: integer
      userID:
        type: string
      username:
        type: string
    type: object
  incluster.PodInfo:
    properties:
      analysisName:
        type: string
      appID:
        type: string
      appName:
        type: string
      containerStatuses:
        items:
          $ref: '#/definitions/v1.ContainerStatus'
        type: array
      creationTimestamp:
        type: string
      externalID:
        type: string
      initContainerStatuses:
        items:
          $ref: '#/definitions/v1.ContainerStatus'
        type: array
      message:
        type: string
      name:
        type: string
      namespace:
        type: string
      phase:
        type: string
      reason:
        type: string
      userID:
        type: string
      username:
        type: string
    type: object
  incluster.ResourceInfo:
    properties:
      configMaps:
        items:
          $ref: '#/definitions/incluster.ConfigMapInfo'
        type: array
      deployments:
        items:
          $ref: '#/definitions/incluster.DeploymentInfo'
        type: array
      pods:
        items:
          $ref: '#/definitions/incluster.PodInfo'
        type: array
      routes:
        items:
          $ref: '#/definitions/incluster.RouteInfo'
        type: array
      services:
        items:
          $ref: '#/definitions/incluster.ServiceInfo'
        type: array
    type: object
...
  incluster.RouteInfo:
    properties:
      analysisName:
        type: string
      appID:
        type: string
      appName:
        type: string
      creationTimestamp:
        type: string
      externalID:
        type: string
      name:
        type: string
      namespace:
        type: string
      rules:
        items:
          $ref: '#/definitions/v1.HTTPRouteRule'
        type: array
      userID:
        type: string
      username:
        type: string
    type: object
  incluster.ServiceInfo:
    properties:
      analysisName:
        type: string
      appID:
        type: string
      appName:
        type: string
      creationTimestamp:
        type: string
      externalID:
        type: string
      name:
        type: string
      namespace:
        type: string
      ports:
        items:
          $ref: '#/definitions/incluster.ServiceInfoPort'
        type: array
      userID:
        type: string
      username:
        type: string
    type: object
  incluster.ServiceInfoPort:
    properties:
      name:
        type: string
      nodePort:
        type: integer
      port:
        type: integer
      protocol:
        type: string
      targetPort:
        type: integer
      targetPortName:
        type: string
    type: object
...
  model.Container:
    properties:
      container_devices:
        items:
          $ref: '#/definitions/model.Device'
        type: array
      container_volumes:
        items:
          $ref: '#/definitions/model.Volume'
        type: array
      container_volumes_from:
        items:
          $ref: '#/definitions/model.VolumesFrom'
        type: array
      cpu_shares:
        type: integer
      entrypoint:
        type: string
      gpu_models:
        description: A list of acceptable GPU models. Empty means no limitation.
        items:
          type: string
        type: array
      id:
        type: string
      image:
        $ref: '#/definitions/model.ContainerImage'
      interactive_apps:
        $ref: '#/definitions/model.InteractiveApps'
      max_cpu_cores:
        description: The maximum number of cores the container needs.
        type: number
      max_gpus:
        description: The maximum number of GPUs the container needs.
        type: integer
      memory_limit:
        description: The maximum the container is allowed to have.
        type: integer
      min_cpu_cores:
        description: The minimum number of cores the container needs.
        type: number
      min_disk_space:
        description: The minimum amount of disk space that the container needs.
        type: integer
      min_gpus:
        description: The minimum number of GPUs the container needs.
        type: integer
      min_memory_limit:
        description: The minimum the container needs.
        type: integer
      name:
        type: string
      network_mode:
        type: string
      pids_limit:
        type: integer
      ports:
        items:
          $ref: '#/definitions/model.Ports'
        type: array
      skip_tmp_mount:
        type: boolean
      uid:
        type: integer
      working_directory:
        type: string
    type: object
  model.ContainerImage:
    properties:
      auth:
        type: string
      id:
        type: string
      name:
        type: string
      osg_image_path:
        type: string
      tag:
        type: string
      url:
        type: string
    type: object
  model.Device:
    properties:
      cgroup_permissions:
        type: string
      container_path:
        type: string
      host_path:
        type: string
    type: object
  model.ExtraInfo:
    properties:
      htcondor:
        $ref: '#/definitions/model.HTCondorExtraInfo'
    type: object
  model.FileMetadata:
    properties:
      attr:
        type: string
      unit:
        type: string
      value:
        type: string
    type: object
  model.HTCondorExtraInfo:
    properties:
      extra_requirements:
        type: string
    type: object
  model.InteractiveApps:
    properties:
      backend_url:
        description: |-
          Only used if you need to override the default backendURL, which should be
          http://<container_name>.
        type: string
      cas_url:
        description: The base URL for the CAS server.
        type: string
      cas_validate:
        description: The path to the validate endpoint on the CAS server.
        type: string
      frontend_url:
        description: |-
          The URL for the frontend of the application. Will get prefixed with the job
          id.
        type: string
      proxy_image:
        description: |-
          The docker image for the reverse proxy that runs on the cluster with the
          job steps.
        type: string
      proxy_name:
        description: The name of the container for the reverse proxy.
        type: string
      ssl_cert_path:
        description: The path to the SSL cert file on the Condor nodes.
        type: string
      ssl_key_path:
        description: The path to the SSL key file on the Condor nodes.
        type: string
      websocket_path:
        description: |-
          If websocket handling requires a special path in the app. The default is to
          have this be empty.
        type: string
      websocket_port:
        description: |-
          If websocket handling requires a special port in the app. The default is to
          use the same port as the backend URL.
        type: string
      websocket_proto:
        description: If websocket handling requires a protocol other than ws://.
        type: string
    type: object
...
  model.Step:
    properties:
      component:
        $ref: '#/definitions/model.StepComponent'
      config:
        $ref: '#/definitions/model.StepConfig'
      environment:
        $ref: '#/definitions/model.StepEnvironment'
      input:
        items:
          $ref: '#/definitions/model.StepInput'
        type: array
      log-file:
        type: string
      output:
        items:
          $ref: '#/definitions/model.StepOutput'
        type: array
      stderr:
        type: string
      stdin:
        type: string
      stdout:
        type: string
      type:
        type: string
    type: object
  model.StepComponent:
    properties:
      container:
        $ref: '#/definitions/model.Container'
      description:
        type: string
      interactive:
        type: boolean
      location:
        type: string
      name:
        type: string
      restricted:
        type: boolean
      time_limit_seconds:
        type: integer
      type:
        type: string
    type: object
  model.StepConfig:
    properties:
      input:
        items:
          $ref: '#/definitions/model.StepInput'
        type: array
      output:
        items:
          $ref: '#/definitions/model.StepOutput'
        type: array
      params:
        items:
          $ref: '#/definitions/model.StepParam'
        type: array
    type: object
  model.StepEnvironment:
    additionalProperties:
      type: string
    type: object
  model.StepInput:
    properties:
      id:
        type: string
      multiplicity:
        type: string
      name:
        type: string
      property:
        type: string
      retain:
        type: boolean
      ticket:
        type: string
      type:
        type: string
      value:
        type: string
    type: object
  model.StepOutput:
    properties:
      multiplicity:
        type: string
      name:
        type: string
      property:
        type: string
      qual-id:
        type: string
      retain:
        type: boolean
      type:
        type: string
    type: object
  model.StepParam:
    properties:
      id:
        type: string
      name:
        type: string
      order:
        type: integer
      path:
        type: string
      type:
        type: string
      value:
        type: string
    type: object
  model.Volume:
    properties:
      container_path:
        type: string
      host_path:
        type: string
      mode:
        type: string
      read_only:
        type: boolean
    type: object
  model.VolumesFrom:
    properties:
      auth:
        type: string
      container_path:
        type: string
      host_path:
        type: string
      name:
        type: string
      name_prefix:
        type: string
      read_only:
        type: boolean
      tag:
        type: string
      url:
        type: string
    type: object
...
host: localhost:60000
info:
  contact: {}
  description: The app-exposer API for the Discovery Environment's VICE feature.
  license:
    name: 3-Clause BSD License
    url: https://github.com/cyverse-de/app-exposer?tab=License-1-ov-file#readme
  title: app-exposer
  version: "1.0"
paths:
  /batch/launch:
    post:
      description: Launch a batch analysis
      operationId: batch-launch
      parameters:
      - description: Analysis Definition
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httphandlers.AnalysisLaunch'
      - default: false
        description: Bypass resource tracking
        in: query
        name: disable-resource-tracking
        type: boolean
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Launch a batch analysis
  /batch/stop/{id}:
    delete:
      description: Stop a batch analysis by its external UUID
      operationId: batch-stop
      parameters:
      - description: External UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Stop a batch analysis by its external UUID
  /cleanup:
    post:
      consumes:
      - application/json
      description: Stop a batch analysis by its external UUID
      operationId: batch-cleanup
      parameters:
      - description: External UUID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httphandlers.uuidBody'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Stop a batch analysis by its external UUID
...
  /vice/{host}/description:
    get:
      description: |-
        Returns a listing entry for a single analysis associated
        with the host/subdomain passed in as 'host' from the URL.
        The user passed in must have access to the VICE analysis.
      operationId: describe-analysis
      parameters:
      - description: username
        in: query
        name: user
        required: true
        type: string
      - description: subdomain
        in: path
        name: host
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/incluster.ResourceInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Returns resources by user and subdomain.
...
  /vice/{id}/download-input-files:
    post:
      description: Triggers input downloads for an analysis
      operationId: trigger-downloads
      parameters:
      - description: External ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Triggers input downloads for an analysis
  /vice/{id}/exit:
    post:
      description: |-
        Terminates the VICE analysis deployment and cleans up
        resources asscociated with it. Does not save outputs first. Uses
        the external-id label to find all of the objects in the configured
        namespace associated with the job. Deletes the following objects:
        HTTP routes, services, deployments, and configmaps.
      operationId: exit
      parameters:
      - description: The external ID of the VICE analysis
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Terminates a VICE analysis
  /vice/{id}/save-and-exit:
    post:
      description: |-
        Handles requests to save the output files in iRODS and then exit.
        The exit portion will only occur if the save operation succeeds. The operation is
        performed inside of a goroutine so that the caller isn't waiting for hours/days for
        output file transfers to complete.
      operationId: save-and-exit
      parameters:
      - description: external ID of the analysis
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Save files and terminate VICE analysis
  /vice/{id}/save-output-files:
    post:
      description: Triggers output uploads from a running analysis.
      operationId: trigger-uploads
      parameters:
      - description: External ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Triggers output uploads from an analysis
...
  /vice/apply-labels:
    post:
      description: |-
        Asynchronously applies labels to all running VICE analyses.
        The application of the labels may not be complete by the time the response is returned.
      operationId: apply-async-labels
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Applies labels to running VICE analyses.
  /vice/async-data:
    get:
      description: |-
        Returns data that is applied to analyses outside of an API call.
        The returned data is not returned asynchronously, despite the name of the call.
      operationId: async-data
      parameters:
      - description: External ID
        in: query
        name: external-id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httphandlers.AsyncData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Returns data that is generated asynchronously from the job launch.
  /vice/launch:
    post:
      consumes:
      - application/json
      description: |-
        The HTTP handler that orchestrates the launching of a VICE analysis inside
        the k8s cluster. This gets passed to the router to be associated with a route. The Job
        is passed in as the body of the request.
      operationId: launch-app
      parameters:
      - description: The request body containing the analysis details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httphandlers.AnalysisLaunch'
      - default: false
        description: Bypass resource tracking
        in: query
        name: disable-resource-tracking
        type: boolean
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Launch a VICE analysis
  /vice/listing:
    get:
      description: |-
        Returns all of the k8s resources associated with a VICE analysis
        but checks permissions to see if the requesting user has permission
        to access the resource. The rest of the query map is used to filter
        resources returned from the handler.
      operationId: filterable-resources
      parameters:
      - description: username
        in: query
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/incluster.ResourceInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Returns resources for a VICE analysis
...
swagger: "2.0"